requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
sqlalchemy==2.0.49
psycopg2-binary==2.9.12
python-dotenv==1.2.1
